Refactor manager structure in Infrastructure layer
